requirements.txt
transformers 
peft 
datasets 
accelerate 
bitsandbytes